refacotr: Restructure package
