+++ app/src/androidTest/java/campus/tech/kakao/map/MainViewUiTest.kt
@@ -8,7 +8,7 @@ import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.espresso.assertion.ViewAssertions.matches
-import campus.tech.kakao.map.view.MainActivity
+import campus.tech.kakao.map.presentation.view.MainActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
