+++ app/src/main/AndroidManifest.xml
@@ -21,16 +21,16 @@
         android:theme="@style/Theme.Map"
         tools:targetApi="31">
         <activity
-            android:name=".view.MapViewActivity"
+            android:name=".presentation.view.MapViewActivity"
             android:exported="false" />
         <activity
-            android:name=".view.ErrorActivity"
+            android:name=".presentation.view.ErrorActivity"
             android:exported="false" />
         <activity
-            android:name=".view.MainActivity"
+            android:name=".presentation.view.MainActivity"
             android:exported="true" />
         <activity
-            android:name=".view.WelcomeActivity"
+            android:name=".presentation.view.WelcomeActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -38,7 +38,7 @@
             </intent-filter>
         </activity>
         <service
-            android:name=".service.FCMservice"
+            android:name=".data.remote.service.FCMservice"
             android:exported="true">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
