+++ app/src/main/java/campus/tech/kakao/map/data/remote/service/FCMservice.kt
@@ -1,0 +1,12 @@
+package campus.tech.kakao.map.data.remote.service
+
+import android.util.Log
+import com.google.firebase.messaging.FirebaseMessagingService
+import com.google.firebase.messaging.RemoteMessage
+
+class FCMservice : FirebaseMessagingService() {
+    override fun onMessageReceived(remoteMessage: RemoteMessage) {
+        super.onMessageReceived(remoteMessage)
+        Log.d("TEST_FcmService", "Received Message")
+    }
+}
