+++ app/src/main/java/campus/tech/kakao/map/data/repository/log/LogRepository.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.repository
+package campus.tech.kakao.map.data.repository.log
 
 import campus.tech.kakao.map.data.db.PlaceDao
 import campus.tech.kakao.map.data.db.entity.Place
