+++ app/src/main/java/campus/tech/kakao/map/data/repository/map/MapRepositoryInterface.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.repository
+package campus.tech.kakao.map.data.repository.map
 
 interface MapRepositoryInterface {
     fun getLastLocation(): Pair<Double, Double>?
