+++ app/src/main/java/campus/tech/kakao/map/presentation/view/MainActivity.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.view
+package campus.tech.kakao.map.presentation.view
 
 import android.content.Intent
 import android.os.Bundle
@@ -8,11 +8,13 @@ import androidx.core.view.isVisible
 import androidx.core.widget.doAfterTextChanged
 import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.LinearLayoutManager
-import campus.tech.kakao.map.viewmodel.PlaceViewModel
+import campus.tech.kakao.map.presentation.viewmodel.PlaceViewModel
 import campus.tech.kakao.map.data.db.entity.Place
 import campus.tech.kakao.map.R
 import campus.tech.kakao.map.databinding.ActivityMainBinding
-import campus.tech.kakao.map.viewmodel.LogViewModel
+import campus.tech.kakao.map.presentation.adapter.RecyclerViewAdapter
+import campus.tech.kakao.map.presentation.adapter.TapViewAdapter
+import campus.tech.kakao.map.presentation.viewmodel.LogViewModel
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
