+++ app/src/main/java/campus/tech/kakao/map/presentation/viewmodel/PlaceViewModel.kt
@@ -1,11 +1,11 @@
-package campus.tech.kakao.map.viewmodel
+package campus.tech.kakao.map.presentation.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import campus.tech.kakao.map.data.db.entity.Place
-import campus.tech.kakao.map.repository.PlaceRepositoryInterface
+import campus.tech.kakao.map.data.repository.place.PlaceRepositoryInterface
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
