--- app/src/main/java/campus/tech/kakao/map/service/FCMservice.kt
@@ -1,69 +1,0 @@
-package campus.tech.kakao.map.service
-
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.app.PendingIntent
-import android.content.Context
-import android.content.Intent
-import android.os.Build
-import android.util.Log
-import androidx.core.app.NotificationCompat
-import campus.tech.kakao.map.R
-import campus.tech.kakao.map.repository.WelcomeRepository
-import campus.tech.kakao.map.view.WelcomeActivity
-import com.google.firebase.messaging.FirebaseMessagingService
-import com.google.firebase.messaging.RemoteMessage
-
-class FCMservice : FirebaseMessagingService() {
-    override fun onMessageReceived(remoteMessage: RemoteMessage) {
-        super.onMessageReceived(remoteMessage)
-
-        Log.d("TEST_FcmService", "Received Message")
-        sendCustomNotification(applicationContext, remoteMessage.data["title"], remoteMessage.data["body"])
-    }
-
-    fun sendCustomNotification(context: Context, title: String?, body: String?) {
-        val intent = Intent(context, WelcomeActivity::class.java).apply {
-            flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-        }
-        val pendingIntent: PendingIntent = PendingIntent.getActivity(context,
-            REQUEST_CODE, intent, PendingIntent.FLAG_IMMUTABLE)
-
-        val notification = NotificationCompat.Builder(context, CHANNEL_ID)
-            .setSmallIcon(R.drawable.icon)
-            .setContentTitle(title)
-            .setContentText(body)
-            .setPriority(NotificationCompat.PRIORITY_DEFAULT)
-            .setContentIntent(pendingIntent)
-            .setAutoCancel(true)
-            .build()
-
-        val notificationManager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-        notificationManager.notify(NOTIFICATION_ID, notification)
-    }
-
-    fun createNotificationChannel(context: Context) {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val channelId = CHANNEL_ID
-            val channelName = CHANNEL_NAME
-            val channelDescription = CHANNEL_DESCRIPTION
-            val importance = NotificationManager.IMPORTANCE_DEFAULT
-
-            val channel = NotificationChannel(channelId, channelName, importance).apply {
-                description = channelDescription
-            }
-
-            val notificationManager: NotificationManager =
-                context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-            notificationManager.createNotificationChannel(channel)
-        }
-    }
-
-    companion object {
-        private const val NOTIFICATION_ID = 222222
-        private const val REQUEST_CODE = 0
-        private const val CHANNEL_ID = "Mychannel"
-        private const val CHANNEL_NAME = "Mychannel"
-        private const val CHANNEL_DESCRIPTION = "This is Mychannel"
-    }
-}
