+++ app/src/main/res/layout/activity_error.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.ErrorActivity">
+    tools:context=".presentation.view.ErrorActivity">
 
     <Space
         android:layout_width="0dp"
