+++ app/src/main/res/layout/activity_map_view.xml
@@ -2,15 +2,15 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".view.MapViewActivity">
+    tools:context=".presentation.view.MapViewActivity">
     <data>
         <variable
             name="view"
-            type="campus.tech.kakao.map.view.MapViewActivity" />
+            type="campus.tech.kakao.map.presentation.view.MapViewActivity" />
 
         <variable
             name="viewModel"
-            type="campus.tech.kakao.map.viewmodel.MapViewModel" />
+            type="campus.tech.kakao.map.presentation.viewmodel.MapViewModel" />
     </data>
     <androidx.coordinatorlayout.widget.CoordinatorLayout
         android:layout_width="match_parent"
