+++ app/src/main/res/layout/bottom_sheet.xml
@@ -4,7 +4,7 @@
     <data>
         <variable
             name="viewModel"
-            type="campus.tech.kakao.map.viewmodel.MapViewModel" />
+            type="campus.tech.kakao.map.presentation.viewmodel.MapViewModel" />
     </data>
 
     <LinearLayout
